% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxMKF.R
\name{CoxMKF}
\alias{CoxMKF}
\title{High Dimensional Mediation Analysis using Multiple Knockoffs}
\usage{
CoxMKF(
  X,
  Y,
  M,
  COV,
  penalty = c("MCP", "lasso", "SCAD"),
  intercept = TRUE,
  q1 = 0.2,
  q2 = 0.1,
  gamma = 0.3,
  n_bootstraps = 25
)
}
\arguments{
\item{X}{a vector of the exposure.}

\item{Y}{a vector of the outcome.}

\item{M}{a matrix of the mediators.}

\item{COV}{optional, a matrix of the potential covariates.}

\item{penalty}{the penalty used in the cox regression.}

\item{intercept}{binary, whether an intercept should be included in the regression M~X, the default value is TRUE.}

\item{q1}{the pre-defined FDR level in the screening step.}

\item{q2}{the pre-defined FDR level in the knockoff procedure.}

\item{gamma}{the pre-defined quantile level.}

\item{n_bootstraps}{the pre-defined number of knockoffs in the AKO.}
}
\value{
A list of the CoxMKF results including the IDs of the screening step, the IDs of the selection step and the estimation for the effects.
}
\description{
The main function of CoxMKF which combines the screening step, knockoff step and estimation step.
}
