% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxMKF.R
\name{knockoff_surv}
\alias{knockoff_surv}
\title{Knockoff for Survival Response}
\usage{
knockoff_surv(
  X,
  Y,
  MS,
  COV,
  penalty = c("MCP", "lasso", "SCAD"),
  q = 0.1,
  gamma = 0.3,
  n_bootstraps = 25
)
}
\arguments{
\item{X}{a vector of the exposure.}

\item{Y}{a vector of the outcome.}

\item{MS}{a matrix of the mediators after the pre-liminary screening.}

\item{COV}{optional, a matrix of the potential covariates.}

\item{penalty}{the penalty used in the cox regression.}

\item{q}{the pre-defined FDR level in the knockoff procedure.}

\item{gamma}{the pre-defined quantile level.}

\item{n_bootstraps}{the pre-defined number of knockoffs in the AKO.}
}
\value{
A vector of mediator IDs selected by AKO.
}
\description{
Knockoff for Survival Response
}
\section{Functions}{
\itemize{
\item \code{knockoff_surv}: Knockoff copies for survival outcome and compute the corresponding statistics.
}}

