% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxMKF.R
\name{screen_FDR}
\alias{screen_FDR}
\title{Screening Based on X->M with FDR control}
\usage{
screen_FDR(X, M, COV, intercept = TRUE, fdr = 0.2)
}
\arguments{
\item{X}{a vector of the exposure.}

\item{M}{a matrix of the mediators.}

\item{COV}{optional, a matrix of the potential covariates.}

\item{intercept}{binary, whether an intercept should be included in the regression M~X, the default value is TRUE.}

\item{fdr}{the pre-defined FDR level in the screening step, the default value is 0.2.}
}
\value{
A list of the screening results including the IDs, pvalues and coefficients.
}
\description{
The first step of CoxMKF, which conducts preliminary screening with the p-values from X to M with FDR control.
}
